backend/requirements.txt
asgiref==3.9.2
celery==5.4.0
Django==5.2.6
django-cors-headers==4.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==23.5
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
future==1.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==5.0.1
django-redis==5.4.0
setuptools==80.9.0
sqlparse==0.5.3
gunicorn==23.0.0
uvicorn[standard]==0.24.0
channels==4.0.0
channels-redis==4.2.0
flower==2.0.1
django-celery-beat==2.8.1
jsonschema==4.25.1
logging==0.4.9.6
# OAuth2 dependencies
authlib==1.3.0
requests==2.31.0
google-auth==2.27.0
